refactor: split reconciliation and alert models
