+++ src/tavily_proxy/reconciliation_engine.rs
@@ -9,6 +9,30 @@ enum ReconciliationOutcome {
     LocalPressure,
 }
 
+struct ReconciliationRunResult {
+    settled: i64,
+    completed: i64,
+    no_adjustment: i64,
+    observed: i64,
+    transport_failure_windows: i64,
+    last_transport_kind: Option<&'static str>,
+    semantic_failure_windows: i64,
+    settled_recent: i64,
+    settled_backlog: i64,
+    upstream_429_retry_windows: i64,
+    local_usage_rate_limit_windows: i64,
+    other_retry_windows: i64,
+    key_backoff_window_count: i64,
+    skipped_by_key_backoff: i64,
+    attempted_candidate_count: i64,
+    budget_exhausted: bool,
+    remote_attempt_limit_reached: bool,
+    max_retry_after_until: Option<i64>,
+    hydrate_ms: i64,
+    first_remote_ms: Option<i64>,
+    remote_ms: i64,
+}
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[doc(hidden)]
 pub enum ClaimedReconciliationRunOutcome {
